+++ src/com/kaylerrenslow/armaDialogCreator/gui/fx/main/popup/newControl/EditCustomControlPopup.java
@@ -8,27 +8,28 @@
  * The software is provided "as is", without warranty of any kind, express or implied, including but not limited to the warranties of merchantability, fitness for a particular purpose and noninfringement. in no event shall the authors or copyright holders be liable for any claim, damages or other liability, whether in an action of contract, tort or otherwise, arising from, out of or in connection with the software or the use or other dealings in the software.
  */
 
-
-package com.kaylerrenslow.armaDialogCreator.gui.fx.main.actions.mainMenu.edit;
+package com.kaylerrenslow.armaDialogCreator.gui.fx.main.popup.newControl;
 
 import com.kaylerrenslow.armaDialogCreator.control.CustomControlClass;
-import com.kaylerrenslow.armaDialogCreator.gui.fx.main.popup.ChooseCustomControlDialog;
-import com.kaylerrenslow.armaDialogCreator.gui.fx.main.popup.newControl.EditCustomControlPopup;
-import javafx.event.ActionEvent;
-import javafx.event.EventHandler;
+import com.kaylerrenslow.armaDialogCreator.main.Lang;
+import org.jetbrains.annotations.NotNull;
 
 /**
- Created by Kayler on 11/13/2016.
- */
-public class EditCustomControlAction implements EventHandler<ActionEvent> {
-	@Override
-	public void handle(ActionEvent event) {
-		ChooseCustomControlDialog dialog = new ChooseCustomControlDialog();
-		dialog.show();
-		CustomControlClass controlClass = dialog.getChosenItem();
-		if (dialog.wasCancelled() || controlClass == null) {
-			return;
-		}
-		new EditCustomControlPopup(controlClass).show();
+ @author Kayler
+ Popup window that allows for editing a {@link CustomControlClass}.
+ Created on 11/14/2016. */
+public class EditCustomControlPopup extends NewControlPopup {
+
+	public EditCustomControlPopup(@NotNull CustomControlClass toEdit) {
+		myStage.setTitle(Lang.ApplicationBundle().getString("Popups.EditCustomControl.popup_title"));
+		setToControlClass(toEdit.getControlClass());
+		disableBaseControlMenuButton(true);
 	}
+
+	@Override
+	protected void ok() {
+
+		close();
+	}
+
 }