+++ CHANGES.md
@@ -2,7 +2,7 @@
 * none
 
 **Changed:**
-* none
+* custom controls can now be edited
 
 **Fixed:**
 * none